LPS-59237 move common logic to a Base class
