+++ modules/apps/application-list/application-list-api/src/com/liferay/application/list/BasePanelApp.java
@@ -1,0 +1,66 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.application.list;
+
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.util.JavaConstants;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.model.Group;
+import com.liferay.portal.model.Portlet;
+import com.liferay.portal.security.permission.PermissionChecker;
+
+import java.util.Locale;
+
+/**
+ * @author Adolfo Pérez
+ */
+public abstract class BasePanelApp implements PanelApp {
+
+	@Override
+	public String getKey() {
+		Class<?> clazz = getClass();
+
+		return clazz.getName();
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		return LanguageUtil.get(
+			locale,
+			JavaConstants.JAVAX_PORTLET_TITLE + StringPool.PERIOD +
+				getPortletId());
+	}
+
+	public Portlet getPortlet() {
+		return _portlet;
+	}
+
+	@Override
+	public boolean hasAccessPermission(
+			PermissionChecker permissionChecker, Group group)
+		throws PortalException {
+
+		return true;
+	}
+
+	@Override
+	public void setPortlet(Portlet portlet) {
+		_portlet = portlet;
+	}
+
+	private Portlet _portlet;
+
+}
